requirements.txt
Flask==2.2.1 
Flask-Login==0.6.2
Jinja2==3.1.2 
flask-compress==1.12
Flask-SQLAlchemy==2.5.1
Flask-WTF==0.14.3
MarkupSafe==2.1.1
pandas==1.3.5
pgeocode==0.3.0
Pillow==9.1.1
psycopg2==2.9.3
pymongo==3.11.0
pyotp==2.6.0
pytest==6.2.4
seaborn==0.11.2
sklearn
statsmodels==0.13.2
Werkzeug==2.2.1
